BREAKING Text.Lexer: rename symbol and space
`symbol` and `space` have been renamed to `symbols` and `spaces` for
consistency, as they recognised sequences.

Two new functions, `symbol` and `space`, have been added as singular
versions of the above.

Previously, the above functions had singular names but plural behavior.
The documentation was also worded in a singular way. This was misleading
and has been corrected.

This breaks existing parsers that rely on these functions, obviously.