Text.Lexer: fix behavior of `reject`
Previously `reject` would fail if the lexer didn't recognise due to
seeing end-of-input right away. This is because internally it used
`Expect` with an inverted predicate. In order to fix this so that
`reject` can succeed as the last recognised lexer, a new Boolean
argument was added to the `Expect` constructor, indicating positive or
negative lookahead. This is processed in `scan`.